cloning an index takes user to newIndex web UI form
Previously, clicking the clone button would lead to a pop-up prompt
box for the name of the cloned index.

With this change the UI instead pre-populates a new index web UI form
with information from the original index, allowing the user to make
tweaks on the parameters of the cloned index.