MB-16714 - allow comma-separated multiple bindHttp args
This allows cbft to optionally listen on multiple addresses.

An example...

  ./cbft -bindHttp=192.168.0.10:9110,0.0.0.0:9110 ...

Only the first address (192.168.0.10) will be used for node cfg
registration, so the other nodes will in the cluster will try to
contact this node as 192.168.0.10:9110.

Because of the second "0.0.0.0:9110" entry, cbft's REST/web-UI will
also be available on localhost.

Change-Id: Ifc8c20981c2dd893f0405f09432caa096475a54f
Reviewed-on: http://review.couchbase.org/56710
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>