MB-16757 - cmd-line option startCheckServer=skip
The cmd-line --option startCheckServer=skip tells cbft to not validate
the --server URL param during startup initializaation, so that the
server URL param can be to a URL that's not actually running yet or
still coming online.

This can be useful in watson ns-server integration, where there might
be startup races as ns-server spawns off cbft.

Change-Id: I6279c6773efc252b162b252a2561db0607496a2b
Reviewed-on: http://review.couchbase.org/58064
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>