MB-17378 - listen on 0.0.0.0's first
With this change, cbft now treats 0.0.0.0 addrs with higher
precendence.  Any other non-0.0.0.0 addr that have the same PORT
number, where cbft is already listening on "0.0.0.0:PORT", will be
logged but treated as a no-op, and should fix the "bind: address
already in use" error cases.

For example, with "127.0.0.1:9910,0.0.0.0:9910", the 0.0.0.0 takes
precedence and the 127.0.0.1 will become a logged no-op.

Change-Id: I4a8938bdb2b6c522ec7cfb3e4ba470ddb3f0d3cf
Reviewed-on: http://review.couchbase.org/58823
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>