MB-17986: remove non-FTS markup and name fts states correctly
In order to integrate with the Angular UI state in a nice way
pluggable components must name their states as substates of the
OOTB under which their plugging in. It's a complexity of
pluging in that, to be fair, isn't well described in the pluggble
UI registry API.

Anyway, with the name change we can now remove the duplicated
markup that had previously been required to make things look like
they were integrated.

Change-Id: I73174054f83cd0619d031066ebdd015fc3364585
Reviewed-on: http://review.couchbase.org/59693
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>