MB-19342 - allow search/count to work when some pindexes missing
use cbgt's new method CoveringPIndexesBestEffort
this gives us a list of pindex names which are currently
unavailable.  we wrap these in a bleve index implementation which
simply reports this as an error.  the rest of the code works as
before, allowing these errors to be part of a partial result set

Change-Id: I71379c93378b2b1f08749378d5f31aa5d1a9b4a8
Reviewed-on: http://review.couchbase.org/63293
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Marty Schoch <marty.schoch@gmail.com>