MB-21474 - optional QueryPIndexes/pindexNames query params
A query can now have an optional pindexNames param, allowing the
client to specify the pindexes that the query handler should consider.

This is useful for a scatter/gathering optimization so that queries
against multiple pindexes on a node would use just a single REST
request per host, instead a separate REST request per pindex/shard.

Change-Id: I3a02cdc85e8becb5651551414461759bedd9872f
Reviewed-on: http://review.couchbase.org/69455
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>