MB-21918 - new option bleveAnalysisQueueSize defaults to numCPU's
The new env option "bleveAnalysisQueueSize" defaults to the number of
cpus.  If you instead specify a positive number, it takes that value
for the bleveAnalysisQueueSize.  If you specify a number X <= 0, it
calculates the "numCPU's - X" to be the bleveAnalysisQueueSize.

As part of this change, other bleve related initialization code is
refactored into the new InitBleveOptions() func.

Change-Id: Ieeb823601ba33a81332b7adcdf886fd0d0383246
Reviewed-on: http://review.couchbase.org/70712
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>