MB-21856 - UI fix to prevent simple cyclic alias definitons
There are 3 cases: creating a brand new index alias, editing an
existing index alias, or cloning an existing index alias.

When editing an existing index alias X, this change adds a check so
that X isn't listed in the multi-select UI control of potential alias
targets, so that it's harder to create an index alias definition that
cyclically points to itself.  This is a UI change only, so defining a
self-cyclic alias with the REST API is still possible.

Additionally, this change only addresses the simplest UI case of alias
X pointing to itself.  The harder cases 2 or more aliases is not
handled.  For example, creating alias X to point to an existing alias
Y and then editing alias Y to point back to alias X is not handled.

The other 2 cases mentioned at the top of creating a new alias or
cloning an exising alias will have all potential alias targets listed.

Change-Id: I68cb21d717d8162ace76b47b5d5d790eae499454
Reviewed-on: http://review.couchbase.org/71069
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>