MB-21858 - spock UI added edit/clone/delete buttons to index view page
I just put the buttons at the bottom of the page, assuming we'll get
somebody with taste to tell us where to really put them later.

The deleteIndex() JS callback looks a lot like the existing
deleteIndex() JS, but is slightly different in that it's aware of
spock's new $state.go() API for changing locations, instead of using
the old $location API.

The CSS adds a little whitespace before the new buttons.

Change-Id: Icc6cee0ea098c0d4d59cd29ae4eafe824c75d4e2
Reviewed-on: http://review.couchbase.org/71070
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>