MB-21861 - spock-UI looks at q, p URL params during searches
Two parts to this change.  First, the angularjs state/router
registration approach that's now used in the new spock UI needs to
know about the q and p params that are part of FTS search URL.
Second, the code was seemingly inconsistent between using 'q' and 'p'
versus 'query' and 'page', so not sure if this ever worked before; and
now, this change now just uses 'q' and 'p' as the param names
consistently.

Change-Id: I82dd60f78dc2a3547d8a4d906dfccee8e18dbe9c
Reviewed-on: http://review.couchbase.org/71077
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>