MB-20831 - spock-UI: hide index def. drill-down info from list page
Rob and Will discussed the changes they wanted to the spock-UI, and
their conclusion / UI spec update was that we should remove the index
definition view from the list page.  Put another way, remove those
sub-expanding triangles.

If users want to see the details of an index definition, they can
either...

A) look at the index definition's JSON.

B) hit the "edit" button to naviate to the index definition's
   edit screen.

For folks who want to understand the motivations and reasoning more, I
defer to Rob and Will.  But, at this point, minimizing the time I
struggle with UI, I just CSS'ed the relevant section to display none,
as there's a non-zero chance that folks will change their minds and
might want this view-related UI back again (such as when a read-only
RBAC user makes a sudden appearance).

Change-Id: I7b782645130addb9455241a853754eeda938d556
Reviewed-on: http://review.couchbase.org/72433
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>