MB-20583 - spock-UI: moved replica dropdown out of adv. subpanel
Moved the Replicas dropdown out of the advanced options subpanel.

Also, fixed an initialization bug with the Replicas dropdown, where it
would incorrectly start off as empty instead of having "0"
preselected.

Moving the Replicas dropdown out to the "top level" made the UI look
(imho) unbalanced, so gave the "index mapping" section an explicit
"Mapping" label to delineate it, and reworded the advanced subpanel
checkbox to have an "Additional options" label, since it was too close
to the label of the "Advanced" expando subpanel from the index mapping
section.

Also, gave the Replicas dropdown a tooltip.

Change-Id: Ibfb544c7cb908c9faa34e1adc3f85e0d8b6216e4
Reviewed-on: http://review.couchbase.org/73993
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>