MB-22937 - spock-UI: revert to simple checkboxes in nested mapping UI
The spock styled checkboxes and associated labels need a unique id per
checkbox in order to work.  But, the FTS mapping UI subpanel is
repeated and nested, so the same id's are incorrectly reused,
resulting in broken checkbox behavior.

Attempts to fix this with dynamically generated unique id's for each
of the recursive ng-include's didn't work, so this change reverts back
to simple checkboxes in the recursive/nested parts of the previous
change dde77bc330382e9023d7b47bffe0bb7657c939ee.

Change-Id: I4d33ff22b12dbd1aaa7053a11e1973131acb5017
Reviewed-on: http://review.couchbase.org/74301
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>