MB-23167 - remap ":PORT" to "localhost:PORT" only for http
In particular, by changing to perform a remapping of ":PORT" only for
http protocol, this allows https endpoints to bind by default to all
addresses.

In addition, decided to keep the remapping for http, as there probably
was some reason to originally do so, perhaps related to a better
security practice by default when cbft is run in standalone mode.

Change-Id: I34fc6b5896c2bf5a8324b850b1a2e355f636c085
Reviewed-on: http://review.couchbase.org/74747
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>