MB-18230 - spock-UI: check if field property+name already taken
In the nested type mapping UI, when the user is adding a new field or
editing an existing field, check if the field's property + name is
already taken by a sibling field.  If already taken, disable the OK
button and style the property label (field) and name label (searchable
as) as red.

Change-Id: If40993d65362a8f41402a5db69532bfbfb869105
Reviewed-on: http://review.couchbase.org/74818
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>