MB-18229 - spock-UI: check if mapping name already taken
In the nested type mapping UI, as the user is creating a new mapping
or is editing an existing mapping, dynamically check if the mapping's
name is already taken by a sibling mapping.  If already taken, disable
the OK button and style the mapping name input control's text as red.

Change-Id: Ic95ebf354c0a48e0fc786abc2e2aa519645a21df
Reviewed-on: http://review.couchbase.org/74819
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>