MB-23196 - REST: filtering of /api/index list results
The FilteredListIndexHandler performs a cbauth.IsAllowed() permission
check and filtering on each of the candidate indexes and aliases that
might be returned.  This new handler is registered onto the mux router
and overrides the /api/index default ListIndexHandler from cbgt.rest.

In rest_perms.go, the /api/index permission is now...

   cluster.bucket{}.fts!read

...where the braces '{}' signal that the response is dynamically
filtered based on the request's user.

Also, declared an explicit permission to an existing REST
stats-related endpoint, which is used during the index progress
percentage computation by the UI...

   /api/stats/sourceStats/{indexName}
   cluster.bucket[<sourceName>].stats.fts!read

Also, the syntax for the permission of access to all buckets needs
empty brackets '[]' in order to work correctly with cbauth.

Also, the CREATE method is removed, simplifying creation and update of
index definitions to just using the PUT method.

Finally, this change also tries to simplify logic around the recursive
sourceNames preparation.

Change-Id: I193c75c62146041728f0b137379c1023dedeae93
Reviewed-on: http://review.couchbase.org/75423
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>