MB-19393 - spock-UI: don't disappear empty DocumentMappings
The bleve-mapping-ui has an initialization step that recursively walks
through the nested DocumentMappings to convert them into a format
that's friendly for the tree-UI control.  In particular, it converts
leaf DocumentMappings that have no child properties in a special way,
promoting their child fields up to the parent mapping as so-called
"propertied fields".  This has the effect of hiding those leaf
DocumentMappings in the tree-UI.

The issue was the check to perform this special propertied-fields case
was too broad, treating DocumentMappings that had zero child fields as
falling into this conversion case... resulting in any empty
DocumentMappings incorrectly disappearing from the tree-UI.

Change-Id: Ia4b24f5be1f85e34c7101a2ea51a55eb408ac48a
Reviewed-on: http://review.couchbase.org/75037
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>