MB-18102 - REST API: improve message on json.UnmarshalTypeError
The error type returned in this specific case by the json Unmarshaler
was a json.UnmarshalTypeError, which cbft now detects via a type cast
attempt.  So, after this change, the output error message now looks
like...

    rest_create_index: error creating index: myFTSIndex,
    err: manager_api: CreateIndex, invalid, err: bleve:
    validate params, JSON parse was expecting a string
    key/field-name but instead saw a bool

Side note, the UnmarshalTypeError has an Offset field, which the
encoding/json API docs seem to hint is the byte offset near the error
in the original JSON...

    https://golang.org/src/encoding/json/decode.go#L121
    Offset int64 // error occurred after reading Offset bytes

...but unfortunately, that Offset seems to be incorrect and
unreliable, circa golang 1.8.

Change-Id: I7ba36e4961c7271c98bb17ed0f1c682ded1fe899
Reviewed-on: http://review.couchbase.org/75841
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>