MB-23674 - do not print *cbgt.PIndex with %#v format without lock
Printing with %#v format will access the mutex and the closed
field, protected by the mutex, so it cannot be done safely
without having acquired the mutext first.  This change now
simply prints the *cbgt.PIndex.Name, and data important for
the error message in a few checks.

This fix assumes that closed is the only feed needing to be
protected by the mutex (which is implied by the struct field
layout conventions)

Race detecor passed in several runs after this change.

Change-Id: I5c76c184e21d17835937c395f28e8c471b0d5677
Reviewed-on: http://review.couchbase.org/76177
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>