MB-30954 - awake the app-herder if memory use drops
Before this commit, the app-herder will (rightly) pause indexing of an
incoming batch if we're over quota.  And, later events can lead to
resumed indexing...

- a search / query requests finishes
- scorch persister makes progress
- scorch file merger makes progress

But, there are other reasons why the process memory used might drop
(i.e., a big GC finishes), so this commit adds logic to ns_server.go
(which is already regularly monitoring process memory used) and when
there's a drop in the process memory used, it notifies the app-herder
(so that the app-herder in turn might resume indexing).

Change-Id: I89681f4f80e73a45fbdecbb79c8b01054d702614
Reviewed-on: http://review.couchbase.org/98465
Reviewed-by: Sreekanth Sivasankaran <sreekanth.sivasankaran@couchbase.com>
Tested-by: Steve Yen <steve.yen@gmail.com>