MB-30954 - goverseer now supports a kick channel for forcing GC
Before this change, the goverseer (when enabled) would wake up on a
repeating interval and recalculate whether the golang SetGCPercent()
needs to be adjusted based on a memory quota.

This change also adds a kickCh to the goverseer, which callers can use
to force an immediate, aggressive SetGCPercent() invocation.

The kickCh is also de-bounced to prevent too many rapid changes to
SetGCPercent() in a short time.  That is, the kickCh might be notified
repeatedly in the middle of the goverseer's time interval cycle, but
only the first kick in the cycle will lead to an immediate
SetGCPercent().

In this commit, the app-herder is also changed to notify the kickCh
(if the kickCh is non-nil) when the app-herder determines we're over
quota for incoming queries or indexing.

Of note, the goverseer is still disabled by default, but can be
enabled via an env variable like CBFT_ENV_OPTIONS=memCheckInterval=5s

Change-Id: Id0edae36c421ad2b81da1e9eefe44011f2c76f55
Reviewed-on: http://review.couchbase.org/98549
Reviewed-by: Sreekanth Sivasankaran <sreekanth.sivasankaran@couchbase.com>
Tested-by: Steve Yen <steve.yen@gmail.com>