MB-30954 - appHerder.onQueryEvent() distinguishes top-level search
In the case of a single node cluster, an incoming search request was
invoking the appHerder's onQueryEvent() callback 7 times...

* 1 time for the incoming request (the "top-level" search request).
* 6 times for the search on each pindex (assuming 6 pindexes).

This commit provides the "depth" of the request handling to the
appHerder, where the top-level, incoming request will be at depth 0,
and the 6 subsequent "child" pindex requests will be at depth 1.

That allows the appHerder to engage its "freebie handling" policy for
the only the top-level, incoming search request.

Also, as part of this change, the various onQueryEvent() callbacks are
flattened, so there are no more app-herder callback closures created
during a search request, which saves a few allocations.

Change-Id: I2eea0b9a16f8d88ff0c62c895987ebb612b3ee38
Reviewed-on: http://review.couchbase.org/98747
Reviewed-by: Sreekanth Sivasankaran <sreekanth.sivasankaran@couchbase.com>
Tested-by: Steve Yen <steve.yen@gmail.com>