Best effort batching to honor BleveMaxOpsPerBatch
This change attempts the merging of batches
to honor the batch size to BleveMaxOpsPerBatch
in a best effort manner. In case of very low
incoming mutation rate, smaller batches are
flushed to bleve on an arbtrary (100ms) timer
basis. This configurable timer value has be
big enough to let batches accumulate and short
enough to let infrequent/random mutations to
get flushed to bleve without much delay.

Change-Id: I4c836996ecd3d049cda0168552e898e9f2556909
Reviewed-on: http://review.couchbase.org/98802
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Sreekanth Sivasankaran <sreekanth.sivasankaran@couchbase.com>