MB-31981: Fix issue with the query-ctl setting of advanced-random
During scatter gather, after the coordinating node decides on
the pindex destinations to hit, the query is distributed across
the nodes with the chosen pindexes. When the sub-query hits
a remote node during the operation, the chosen pindex name is
cross referenced with the local pindexes on that node. This
means that the partition selection in that node needs to be set
to "advanced-local" to fetch all the pindexes on the node.

Change-Id: I61c500de6d92278f4fda6f5d33874494a89f1339
Reviewed-on: http://review.couchbase.org/103567
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Abhinav Dangeti <abhinav@couchbase.com>