text-always dynamic indexing via optional, 'disabled' DateTimeParser
In dynamic indexing, text fields are first passed through the default
date time parser to see if the text parses as a date string (e.g.,
"2019-01-31").  The usual configured default date time parser of
"dateTimeOptional" has this behavior.

However, if the applicaiton instead wants dynamic indexing to always
treat field values as text, this "dateTimeOptional" behavior does not
work correctly.  For example, the user wants "2019-01-31" to be
treated as text (e.g., perhaps via the keyword analyzer), not indexed
as a date-time value.

This commit registers a new DateTimeParser named "disabled", which
always considers any input string as not-a-date.  When the default
date time parser is configured to be "disabled", then dynamic indexing
will always treat string values as text.

Change-Id: Icb759b8b28e996ac16d1eca2cc11db08310dcfb3
Reviewed-on: http://review.couchbase.org/107102
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Yen <steve.yen@gmail.com>
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>