MB-36166 - FTS Stat for DCP Batches Blocked is unreadable
While creating an index, the FTS page is showing a very large number
for DCP Batches Blocked: 18,446,744,073,709,552,000. There are two
parts to the fix for this. The first is a small function to optionally
format such large numbers in scientific notation. This second change
uses the scientific notation formatting in the FTS page.

Change-Id: I6eb009a2f929a5f00221df90f3a6cac986aca074
Reviewed-on: http://review.couchbase.org/116886
Tested-by: Eben Haber <eben@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>