MB-25475: prepare FTS service to lazy loading
In order to prepare it to lazy loading.

This change introduces new main.js module which
is a root module of FTS UI and place for
FTS UI inital config.

In this change, we also indicate that the main.js
will be imported into generated pluggable-uis.js
(more here http://review.couchbase.org/c/ns_server/+/125551)

Change-Id: I39b96ad12626aafd8522ba8cc77296b817cc1f7a
Reviewed-on: http://review.couchbase.org/c/cbft/+/125786
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>
Tested-by: Pavel Blagodov <stochmail@gmail.com>