MB-40602 - FTS Scrambled query field in response with docID query
With multi collection indexes, document IDs specified in the
docID query gets prefixed with collectionUUID bytes to match
the actual doc IDs with bleve. But this should work transparent
to the end user and the query should be kept intact while
sending the response back to the user.

Change-Id: I87fbf1ffae190e399162a93d4a0ee968f8c01b2f
Reviewed-on: http://review.couchbase.org/c/cbft/+/133119
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>
Tested-by: Sreekanth Sivasankaran <sreekanth.sivasankaran@couchbase.com>