MB-41953: update easy_mode_hash to detect more changes
specifically, now we also consider the doc_config
settings within params as a part of the hash
calculation.

store (in params) and planParams (outside params) are
not considered, but we believe this is safe as the
easy mode editor does not use or change them,
the original values are always preserved.

Change-Id: I11e52df4389379e7976a2ce4a714984db4e293dd
Reviewed-on: http://review.couchbase.org/c/cbft/+/138192
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>
Tested-by: Marty Schoch <marty.schoch@gmail.com>