MB-42186: fix tracking of arrays when parsing
we use the valsSincePop variable to track the case where
we just pushed and are immediately poppping (empty array
or obj) as in those cases the JSON formatter places
the closing brace on the same line.

the original logic happened to work in some simple test
cases, but a sample JSON was provided which exposed it's
limitations.  it is believe this change gives more robust
handling of this situation.

Change-Id: I5b755457c8858dcf3ff94c0fa07be763f62b7ec1
Reviewed-on: http://review.couchbase.org/c/cbft/+/138566
Tested-by: Marty Schoch <marty.schoch@gmail.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>