MB-42651: fix handling of searchable as
Two issues addressed:

Nested fields now correctly show the inherited prefix, and only
allow changing the final portion of the searchable as.  The
table also includes this prefix in the displayed searchable as
column.

When loading the editor state from an existing mapping, the
fields name was incorrectly used as the index, when it should
have used the path.  This was found because the chagne above
meant that path != name.  This problem has been corrected.

Change-Id: Ib852adf4eb6f8e0d65ab67d2c36f53c751eda23f
Reviewed-on: http://review.couchbase.org/c/cbft/+/140062
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>
Tested-by: Abhinav Dangeti <abhinav@couchbase.com>