MB-44446 set timeout for scorch bolt config
If scorch open is blocked for more than 30s, it will now return
an error.  Currently, this will be logged, but no other specific
error handling for this case is added.

The value of 30s is arbitrary, but a somewhat higher
value was chosen to avoid positives caused by slower operation
on a heavily loaded system.

Change-Id: I5852c8d37065579a27ee79f5048d98679c43340b
Reviewed-on: http://review.couchbase.org/c/cbft/+/146999
Tested-by: Marty Schoch <marty.schoch@gmail.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>
Reviewed-by: Sreekanth Sivasankaran <sreekanth.sivasankaran@couchbase.com>