MB-44703 - fts swap rebalance reports progress > 100%
This happens due to incorrect calculation of the number
of partitions getting moved during a rebalance. And that
stems from the lack of knowledge about the number of nodes
in a pre-rebalance cluster. This could be fixed by
deferring the self-registeration of a node to Cfg during
boot up.
This helps the ctl to capture the nodes in the cluster
before rebalance and identify a node addition especially
when the addition happens to be the current node.
(which is also the rebalance orchestrator node)

Change-Id: Ief3b65fba51633978ed873b8f9a09cbd30fb348c
Reviewed-on: http://review.couchbase.org/c/cbft/+/148023
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>
Tested-by: Sreekanth Sivasankaran <sreekanth.sivasankaran@couchbase.com>