MB-52030 - enforce cgroup memory limits in FTS
This PR checks the FTS memory quota option by comparing it
with the cgroup memory limits, if running in a container.
If the FTS memory quota is greater than the cgroup limit,
it is changed to the cgroup limit.

Verification: In a cluster built using a container built from image
build-docker.couchbase.com/couchbase/server-toy:aditi.ahuja-enterprise-7.1.1-14989,
512 MB were allocated for FTS and the container had 8 cores.
On firing the /api/manager endpoint, the response contained the following snippet:
`"ftsCpuQuota":"8","ftsMemoryQuota":"524288000"`.
These values tie in with the expected values of 8 cores and 512 MB.

Change-Id: Ia67db9f96ede8c1c59dccdbf47500f01301e9c77
Reviewed-on: https://review.couchbase.org/c/cbft/+/175658
Well-Formed: Restriction Checker
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>
Tested-by: Aditi Ahuja <aditi.ahuja@couchbase.com>