MB-53962: Using SIGAR to get numCPU
The existing method of using runtime.NumCPU() for
the number of CPUs being used is inaccurate in cases
of the number of CPUs to be used being changed in a
container.
To work around this, when run in a container, SIGAR
is used to obtain the number of CPUs being used.

Change-Id: I456c95b243e29f6474adbd9693c200c503b259ba
Reviewed-on: https://review.couchbase.org/c/cbft/+/181354
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: <thejas.orkombu@couchbase.com>
Reviewed-by: Abhi Dangeti <abhinav@couchbase.com>
Tested-by: Aditi Ahuja <aditi.ahuja@couchbase.com>