MB-54303 : Failure modes on service crash
This PR solves for some failure cases:
- Crash after download complete: adds a check to
distinguish between pindexes that are being
unhibernated after a crash vs in the same run.
This is to avoid re-opening an already open bleve
index.
- Crash during download: avoid downloading files
if the restart is after a crashed download.

Also adds a function to track bucket state during
hibernation and delete/reset indexes based on it.
Adds initialisation of an option to track the
hibernation bucket.

Change-Id: I158f21673b51327864c4a10be20f336e691b1605
Reviewed-on: https://review.couchbase.org/c/cbft/+/183669
Tested-by: Abhi Dangeti <abhinav@couchbase.com>
Reviewed-by: Abhi Dangeti <abhinav@couchbase.com>
Reviewed-by: <thejas.orkombu@couchbase.com>