MB-55127 - [FTS] Metering overflows observed during perf tests
- Also includes the changes for removing the capping mechanism
as per MB-54505.
- The main cause of the problem is that the history
tracked using prevRBytes was incorrect, which caused the incorrect
resetting of the prevBytes value and the immediate next metering op
would now try to meter the cumulative bytesRead as RUs causing an
overflow.
- As per logs, the delta (which is the old way of getting
bytes read for that particular query) is not calculated properly
especially when there are multiple parallel queries, mainly because of
prevRBytes.
- Using the bytesRead returned as part of search response for a partition
now looks more apt and also solves the overflow issue.

Change-Id: I21f9d59363fb0445cff5e29fc192a53f4f91ffbd
Reviewed-on: https://review.couchbase.org/c/cbft/+/182813
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Abhi Dangeti <abhinav@couchbase.com>
Tested-by: <thejas.orkombu@couchbase.com>