MB-62841 - Set node priority in NodeInfo
This PR sets the cluster compatible version
as the priority of a node on cbft start. This
will be used as the Priority when ns_server
fetches the node info of a cluster.

A node with higher priority will be picked as
the rebalance orchestrator, unless being
removed or failed over. Hence, rebalance will
select the more recent version node as rebalance
orchestrator, allowing for rebalance path
improvements to be picked up in the first
rebalance operation during an upgrade.

Change-Id: Ife80451403e86ba177425b101de9c91e318f5f0f
Reviewed-on: https://review.couchbase.org/c/cbft/+/213690
Well-Formed: Restriction Checker
Reviewed-by: <thejas.orkombu@couchbase.com>
Reviewed-by: Abhi Dangeti <abhinav@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Aditi Ahuja <aditi.ahuja@couchbase.com>