MB-62182: Introduce a faster way of merging the data segments           in a vector index
- First comes the sampling phase, where a faiss index is created
using the vectors randomly sampled from KV. The coarse quantizer of
this 'centroid index' is used in the later parts of the
index lifecycle.
- The data ingestion on all the bleveDests is blocked until this
centroid index is generated and streamed to all the partitions on that
node.
- After the sampling phase, the feeds are unlocked at the batch worker
level to ingest the data, which follows the existing codepath.
- https://github.com/blevesearch/bleve/pull/2204

Change-Id: I1f7e35d6b519a4cf5a502a7246f407c33c9dfae5
Reviewed-on: https://review.couchbase.org/c/cbft/+/235240
Tested-by: <thejas.orkombu@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: <thejas.orkombu@couchbase.com>